+++ src/index.html
@@ -22,8 +22,8 @@
       <include src="./partials/pagination.html"></include>
       <include src="./partials/button-scroll.html"></include>
     </main>
-    <!-- <include src="./partials/button-scroll.html"></include> -->
-    <include src="./partials/footer.html"></include>
+
+    <include src="./partials/page-footer.html"></include>
 
     <script src="./index.js"></script>
   </body>
