+++ src/index.js
@@ -4,15 +4,13 @@ import 'basiclightbox/dist/basicLightbox.min.css';
 import '@splidejs/splide/dist/css/themes/splide-sea-green.min.css';
 import './sass/main.scss';
 
-import './js/components/modal-movie';
+import './js/components/render-trending-movies';
+import './js/components/render-search-movies';
+import './js/components/notiflix-options';
 import './js/components/refs';
 
 import './js/components/page-header';
-import './js/components/notiflix-options';
-
-import './js/api/trending-movie-search';
-import './js/components/render-search-movies';
-
-import './js/components/modal-team';
 import './js/components/pagination';
 import './js/components/scrollTop';
+import './js/components/modal-movie';
+import './js/components/modal-team';
