+++ src/js/components/modal-team.js
@@ -1,27 +1,11 @@
 import * as basicLightbox from 'basiclightbox';
 import Splide from '@splidejs/splide';
 import refs from './refs';
-import teamCardsTemplate from '../../templates/team-markup.hbs';
 import team from '../db/team.json';
+import teamSplideOptions from './splide-options';
+import teamCardsTemplate from '../../templates/team-markup.hbs';
 
 let teamLightbox = null;
-const teamSplideOptions = {
-  perPage: 3,
-  breakpoints: {
-    767: {
-      perPage: 1,
-    },
-    1023: {
-      perPage: 2,
-    },
-  },
-  focus: 'left',
-  gap: '30px',
-  speed: 700,
-  // type: 'loop',
-  // autoplay: true,
-  // interval: 3000,
-};
 
 refs.footerLink.addEventListener('click', openTeamLightbox);
