+++ src/js/components/render-search-movies.js
@@ -4,79 +4,40 @@ import { searchApiService } from '../api/apiServicePlugin';
 import imageCardTpl from '../../templates/card-markup.hbs';
 import dataStorage from './data-storage';
 import refs from './refs';
-import movePageOne from './pagination.js';
+import { resetPaginationPage, setTotalItems } from './pagination.js';
 
 refs.sectionHome.addEventListener('submit', onSearch);
 
-// === ВРЕМЕННО, до подключения пагинации - Intersection Observer ===
-const options = {
-  root: null,
-  rootMargin: '0px',
-  threshold: 0.9,
-};
-
-const onEntry = (elements, observer) => {
-  elements.forEach(element => {
-    if (element.isIntersecting) {
-      onLoadMore();
-    }
-  });
-};
-
-const observer = new IntersectionObserver(onEntry, options);
-
-function setObserver() {
-  observer.observe(refs.galleryContainer.lastElementChild);
-}
-
-function removeObserver(data) {
-  observer.unobserve(refs.galleryContainer.lastElementChild);
-  return data;
-}
-// ==============================================================
-
-// --- Функции рендеринга изображений ---
-function onSearch(e) {
+export function onSearch(e) {
   e.preventDefault();
-  refs.galleryContainer.innerHTML = '';
+
   refs.pagination.dataset.pagin = 'input';
-  movePageOne();
+  resetPaginationPage();
   searchApiService.resetPage();
   searchApiService.searchQuery = e.currentTarget.firstElementChild.value.trim();
 
-  if (searchApiService.searchQuery.length < 1) {
+  if (!searchApiService.searchQuery.length) {
     refs.galleryContainer.innerHTML = '';
-    onFetchError();
     e.target.value = '';
+    onFetchError();
     return;
   }
 
+  loadSearchedMovies();
+}
+
+export default function loadSearchedMovies() {
+  refs.galleryContainer.innerHTML = '';
   Loading.circle('Loading...');
 
   searchApiService
     .fetchArticles()
     .then(({ results, total_results }) => {
+      setTotalItems(total_results);
       const currentPageMovies = dataStorage.getFilmData(results);
       dataStorage.saveCurrentMovies(currentPageMovies);
       createGallery(currentPageMovies);
     })
-    .then(setObserver)
-    .catch(onFetchError)
-    .finally(Loading.remove(200));
-}
-
-function onLoadMore() {
-  searchApiService.incrementPage();
-  removeObserver();
-  searchApiService
-    .fetchArticles()
-    .then(removeObserver)
-    .then(data => {
-      const currentPageMovies = dataStorage.getFilmData(data);
-      dataStorage.saveCurrentMovies(currentPageMovies);
-      createGallery(currentPageMovies);
-    })
-    .then(setObserver)
     .catch(onFetchError)
     .finally(Loading.remove(200));
 }
@@ -85,6 +46,6 @@ function createGallery(data) {
   refs.galleryContainer.insertAdjacentHTML('beforeend', imageCardTpl(data));
 }
 
-function onFetchError(message) {
+function onFetchError() {
   Notify.failure('Search result not successful. Enter the correct movie name');
 }
