+++ src/js/components/splide-options.js
@@ -1,0 +1,17 @@
+export default {
+  perPage: 3,
+  breakpoints: {
+    767: {
+      perPage: 1,
+    },
+    1023: {
+      perPage: 2,
+    },
+  },
+  focus: 'left',
+  gap: '30px',
+  speed: 700,
+  // type: 'loop',
+  // autoplay: true,
+  // interval: 3000,
+};
