--- src/partials/buttons.html
@@ -1,5 +1,0 @@
-<!-- Кнопки MODAL -->
-<button class="btn btn--secondary" type="button" data-action="add-to-watched">
-  ADD TO WATCHED
-</button>
-<button class="btn btn--secondary" type="button" data-action="add-to-queue">ADD TO QUEUE</button>
